// src/app/api-client/model/usuario.ts
/**
 * Blog API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Comment } from './comment';
import { Post } from './post';


export interface Usuario { 
    id?: number;
    nome?: string;
    email?: string;
    senha?: string;
    papel?: Usuario.PapelEnum;
    descricao?: string;
    posts?: Array<Post>;
    comentarios?: Array<Comment>;
}
export namespace Usuario {
    export type PapelEnum = 'ADMIN' | 'USER';
    export const PapelEnum = {
        Admin: 'ADMIN' as PapelEnum,
        User: 'USER' as PapelEnum
    };
}